migrated all client/src/components/archive files from jsx to tsx
